+++ CHANGELOG.md
@@ -18,6 +18,7 @@ changes.
 
 - Make testIds for link and identity references in drep form unique [Issue 1928](https://github.com/IntersectMBO/govtool/issues/1928)
 - Fix saving the Do Not List checkbox value on DRep registration [Issue 1940](https://github.com/IntersectMBO/govtool/issues/1940)
+- Fix passing DRep votes to the modal showing vote numbers [Issue 1941](https://github.com/IntersectMBO/govtool/issues/1941)
 
 ### Changed